requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
typing-extensions==4.12.2

# ── LLM & Agents ──────────────────────────────────────────────────────────────
openai==1.30.5
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
langgraph==0.1.1

# ── RAG / Vector Store ────────────────────────────────────────────────────────
faiss-cpu==1.8.0
tiktoken==0.7.0

# ── API clients ───────────────────────────────────────────────────────────────
requests==2.32.3
httpx==0.27.0

# ── UI ────────────────────────────────────────────────────────────────────────
streamlit==1.35.0
Pillow==10.3.0

# ── Utilities ─────────────────────────────────────────────────────────────────
numpy==1.26.4

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-cov==5.0.0


# ── Production ────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
redis==5.0.4
sentry-sdk==2.3.1
prometheus-client==0.20.0
ruff==0.4.4
